Shifted Routes
all frontend routes shifted except news
