+++ routes/projects.js
@@ -153,36 +153,7 @@ router.route('/view_all').get(adminAuth, (req, res) => {
         res.redirect('/admin/profile')      })
 })
 
-router.route('/front/search').get(async (req,res)=>{
-  
-  if(Object.keys(req.query).length==0){
-    projectsModel.find().limit(30)
-    .then(project=>res.json(project))
-    .catch(err=>res.json(err))
-  }
-  else{
-    
-    let branch_name = req.query['branch'] 
-    let club_name = req.query['club'] 
-    let deg_name = req.query['degree'] 
-    let query_string = req.query['query_string']
-    let filters={
-      branch:  branch_name==undefined? /.*/ : branch_name,
-      club: club_name==undefined? /.*/ : club_name,
-      degree: deg_name==undefined? /.*/ : deg_name,
-      $text: {$search: query_string}
-     }
-    query_string === undefined && delete filters["$text"]
-    let projects
-    try {
-      projects = await projectsModel.find(filters,{score:{$meta:'textScore'}}).limit(30).sort({score:{$meta:'textScore'}})
-      res.json(projects)
-    } 
-    catch (error) {
-    res.json(error)
-    }
-  }
-})
+
 
 
 module.exports = router
